[Tests] UPdate thresholding tests
